Slightly Large update
Updated fxmanifest to say PF Team instead of ribbitpoison.
Made it so personal vehicle blip disappears if your in that vehicle.
Updated the about menu so it says PF-vMenu version and added a download link.
Made it so when you set a personal vehicle the menu shows up in the main menu along side its default location.