Refactors weather & time perms check to use PermissionsManager.IsAllowed
What was here and written by poco8537 was bang on the money, but could be simplified by using the existing `PermissionsManager` class and it's existing funcs and enums - which is what this commit refactors.

Also adds a missing permission check for randomizing clouds, and an explicit check for setting time (versus simply having access to the time menu).